Refactor session components, bang head against datepicker wall
